+++ Example/Tests/AttributesCreationTests.swift
@@ -30,47 +30,47 @@ class AttributesCreation: QuickSpec {
             }
             
             it("is initialized with max display priority") {
-                attributes.displayPriority = .max
-                expect(attributes.displayPriority).to(equal(.max))
-                expect(attributes.displayPriority.rawValue).to(equal(EKAttributes.DisplayPriority.maxRawValue))
+                attributes.precedence.priority = .max
+                expect(attributes.precedence.priority).to(equal(.max))
+                expect(attributes.precedence.priority.rawValue).to(equal(EKAttributes.Precedence.Priority.maxRawValue))
             }
             
             it("is initialized with high display priority") {
-                attributes.displayPriority = .high
-                expect(attributes.displayPriority).to(equal(.high))
-                expect(attributes.displayPriority.rawValue).to(equal(EKAttributes.DisplayPriority.highRawValue))
+                attributes.precedence.priority = .high
+                expect(attributes.precedence.priority).to(equal(.high))
+                expect(attributes.precedence.priority.rawValue).to(equal(EKAttributes.Precedence.Priority.highRawValue))
             }
             
             it("is initialized with high display priority") {
-                attributes.displayPriority = .normal
-                expect(attributes.displayPriority).to(equal(.normal))
-                expect(attributes.displayPriority.rawValue).to(equal(EKAttributes.DisplayPriority.normalRawValue))
+                attributes.precedence.priority = .normal
+                expect(attributes.precedence.priority).to(equal(.normal))
+                expect(attributes.precedence.priority.rawValue).to(equal(EKAttributes.Precedence.Priority.normalRawValue))
             }
             
             it("is initialized with low display priority") {
-                attributes.displayPriority = .low
-                expect(attributes.displayPriority).to(equal(.low))
-                expect(attributes.displayPriority.rawValue).to(equal(EKAttributes.DisplayPriority.lowRawValue))
+                attributes.precedence.priority = .low
+                expect(attributes.precedence.priority).to(equal(.low))
+                expect(attributes.precedence.priority.rawValue).to(equal(EKAttributes.Precedence.Priority.lowRawValue))
             }
             
             it("is initialized with min display priority") {
-                attributes.displayPriority = .min
-                expect(attributes.displayPriority).to(equal(.min))
-                expect(attributes.displayPriority.rawValue).to(equal(EKAttributes.DisplayPriority.minRawValue))
+                attributes.precedence.priority = .min
+                expect(attributes.precedence.priority).to(equal(.min))
+                expect(attributes.precedence.priority.rawValue).to(equal(EKAttributes.Precedence.Priority.minRawValue))
             }
             
             it("is initialized with custom display priority") {
-                let custom1 = EKAttributes.DisplayPriority(rawValue: 999)
-                attributes.displayPriority = custom1
-                expect(attributes.displayPriority).to(equal(custom1))
-                expect(attributes.displayPriority.rawValue).to(equal(999))
+                let custom1 = EKAttributes.Precedence.override(priority: .init(999), dropEnqueuedEntries: true)
+                attributes.precedence.priority = custom1.priority
+                expect(attributes.precedence.priority).to(equal(custom1.priority))
+                expect(attributes.precedence.priority.rawValue).to(equal(999))
                 
-                let custom2 = EKAttributes.DisplayPriority(1)
-                attributes.displayPriority = custom2
-                expect(attributes.displayPriority).to(equal(custom2))
-                expect(attributes.displayPriority.rawValue).to(equal(1))
+                let custom2 = EKAttributes.Precedence.override(priority: .init(1), dropEnqueuedEntries: true)
+                attributes.precedence.priority = custom2.priority
+                expect(attributes.precedence.priority).to(equal(custom2.priority))
+                expect(attributes.precedence.priority.rawValue).to(equal(1))
                 
-                expect(custom2).to(beLessThan(custom1))
+                expect(custom2.priority).to(beLessThan(custom1.priority))
             }
         }
     }