+++ Source/Model/EntryAttributes/EKAttributes.swift
@@ -27,8 +27,8 @@ public struct EKAttributes {
     /** The position of the entry inside the screen */
     public var position = Position.top
 
-    /** The display priority of the entry. *.max* is the highest. *.min* is lowest */
-    public var displayPriority = DisplayPriority.normal
+    /** The display manner of the entry. */
+    public var precedence = Precedence.override(priority: .normal, dropEnqueuedEntries: false)
     
     /** Describes how long the entry is displayed before it is dismissed */
     public var displayDuration: DisplayDuration = 2 // Use .infinity for infinite duration