+++ Source/SwiftEntryKit.swift
@@ -8,13 +8,36 @@
 import Foundation
 
 /**
- A stateless, threadsafe entry point that contains the display and the dismissal logic of entries.
+ A stateless, threadsafe (unless described otherwise) entry point that contains the display and the dismissal logic of entries.
  */
 public final class SwiftEntryKit {
     
+    /** Describes the a single or multiple entries for possible dismissal states */
+    public enum EntryDismissalDescriptor {
+        
+        /** Describes specific entry / entries with name */
+        case specific(entryName: String)
+        
+        /** Describes a group of entries with lower or equal display priority */
+        case prioritizedLowerOrEqualTo(priority: EKAttributes.Precedence.Priority)
+        
+        /** Describes all the entries that are currently in the queue and pending presentation */
+        case enqueued
+        
+        /** Describes all the entries */
+        case all
+        
+        /** Describes the currently displayed entry */
+        case displayed
+    }
+    
     /** The window to rollback to after dismissal */
     public enum RollbackWindow {
+        
+        /** The main window */
         case main
+        
+        /** A given custom window */
         case custom(window: UIWindow)
     }
     
@@ -26,7 +49,7 @@ public final class SwiftEntryKit {
     
     /**
      Returns true if **any** entry is currently displayed.
-     - Not a thread safe - should be called from the main queue only in order to receive a reliable result.
+     - Not thread safe - should be called from the main queue only in order to receive a reliable result.
      - Convenience computed variable. Using it is the same as invoking **isCurrentlyDisplaying() -> Bool** (witohut the name of the entry).
      */
     public class var isCurrentlyDisplaying: Bool {
@@ -35,12 +58,32 @@ public final class SwiftEntryKit {
     
     /**
      Returns true if an entry with a given name is currently displayed.
-     - Not a thread safe - should be called from the main queue only in order to receive a reliable result.
+     - Not thread safe - should be called from the main queue only in order to receive a reliable result.
      - If invoked with *name* = *nil* or without the parameter value, it will return *true* if **any** entry is currently displayed.
+     - Returns a *false* value for currently enqueued entries.
      - parameter name: The name of the entry. Its default value is *nil*.
      */
     public class func isCurrentlyDisplaying(entryNamed name: String? = nil) -> Bool {
         return EKWindowProvider.shared.isCurrentlyDisplaying(entryNamed: name)
+    }
+    
+    /**
+     Returns true if **any** entry is currently enqueued and waiting to be displayed.
+     - Not thread safe - should be called from the main queue only in order to receive a reliable result.
+     - Convenience computed variable. Using it is the same as invoking **~queueContains() -> Bool** (witohut the name of the entry)
+     */
+    public class var isQueueEmpty: Bool {
+        return !queueContains()
+    }
+    
+    /**
+     Returns true if an entry with a given name is currently enqueued and waiting to be displayed.
+     - Not thread safe - should be called from the main queue only in order to receive a reliable result.
+     - If invoked with *name* = *nil* or without the parameter value, it will return *true* if **any** entry is currently displayed, meaning, the queue is not currently empty.
+     - parameter name: The name of the entry. Its default value is *nil*.
+     */
+    public class func queueContains(entryNamed name: String? = nil) -> Bool {
+        return EKWindowProvider.shared.queueContains(entryNamed: name)
     }
     
     /**
@@ -90,11 +133,12 @@ public final class SwiftEntryKit {
      Dismisses the currently presented entry and removes the presented window instance after the exit animation is concluded.
      - A thread-safe method - Can be invoked from any thread.
      - A class method - Should be called on the class.
+     - parameter descriptor: A descriptor for the entries that are to be dismissed. The default value is *.displayed*.
      - parameter completion: A completion handler that is to be called right after the entry is dismissed (After the animation is concluded).
      */
-    public class func dismiss(with completion: DismissCompletionHandler? = nil) {
+    public class func dismiss(_ descriptor: EntryDismissalDescriptor = .displayed, with completion: DismissCompletionHandler? = nil) {
         DispatchQueue.main.async {
-            EKWindowProvider.shared.dismiss(with: completion)
+            EKWindowProvider.shared.dismiss(descriptor, with: completion)
         }
     }